+++ apimocking-main/apimocking-boilerplate/src/main/java/com/grepp/spring/app/model/budget/model/BudgetCategorySummary.java
@@ -1,5 +1,6 @@
 package com.grepp.spring.app.model.budget.model;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,5 +14,5 @@ public class BudgetCategorySummary {
 
     private String category;
     private String categoryIcon;
-    private int totalAmount;
+    private BigDecimal totalAmount;
 }
