+++ apimocking-main/apimocking-boilerplate/src/main/java/com/grepp/spring/app/model/budget/model/BudgetGoal.java
@@ -1,6 +1,8 @@
 package com.grepp.spring.app.model.budget.model;
 
 
+import com.google.type.Decimal;
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,6 +16,6 @@ public class BudgetGoal {
 
     private String itemName;
     private String itemImage;
-    private int itemPrice;
+    private BigDecimal itemPrice;
 
 }
